refactor: reorganize validation imports and introduce new validation pipelines names
- Updated import paths for validation-related functions to reflect new module structure, enhancing code organization and maintainability.
- Introduced new validation pipelines for tailor setup and YAML file validation, improving the clarity and functionality of the validation process.
- Added a new path resolution module to handle path validation and resolution logic, ensuring better separation of concerns within the validation framework.
