+++ scripts/validate-schema.ts
@@ -1,6 +1,6 @@
 import { pipe } from 'remeda';
 import { match } from 'ts-pattern';
-import { validateTailorFilesPipeline } from '@shared/validation/validation-only-pipeline';
+import { validateTailorFilesPipeline } from '@shared/validation/tailor-check-pipeline';
 import type { ValidationType, PathResolutionInput } from '@shared/validation/types';
 import { loggers } from '@shared/core/logger';
 import { parseCliArgs } from '@shared/cli/cli-args';
